backend_fastapi/requirements.txt
mysql-connector-python
fastapi
uvicorn
python-multipart
configparser